requirements.txt
google-adk
pandas
matplotlib
python-dotenv